More consistent naming.
Move implementation specifics to impl4 (and not impl3 as this Go implementation
is in fact somewhat different from the impl3 in C++).

Try to be consistent with detection element ID variable name : use the shorter
 one = deid.
