+++ main.go
@@ -2,6 +2,8 @@ package main
 
 import (
 	"github.com/aphecetche/pigiron/mapping"
+	// must include the specific implementation package of the mapping
+	_ "github.com/aphecetche/pigiron/mapping/impl4"
 )
 
 func main() {
