+++ mapping/cathodesegmentation.go
@@ -1,6 +1,7 @@
 package mapping
 
 import (
+	"errors"
 	"fmt"
 	"io"
 	"log"
@@ -8,6 +9,9 @@ import (
 
 	"github.com/aphecetche/pigiron/geo"
 )
+
+// ErrInvalidPadUID signals that an invalid pad uid was used / returned
+var ErrInvalidPadUID = errors.New("invalid pad uid")
 
 type PadUID int
 type DualSampaID int
@@ -31,13 +35,12 @@ type CathodeSegmentation interface {
 	PadSizeY(paduid PadUID) float64
 	GetNeighbours(paduid PadUID) []PadUID
 	IsBending() bool
-	setDetElemID(de int)
 }
 
 // ForEachDetectionElement loops over all detection elements and call the detElemIdHandler function
 // for each of them
-func ForEachDetectionElement(detElemIDHandler func(detElemID int)) {
-	for _, detElemID := range []int{100, 101, 102, 103, 200, 201, 202, 203, 300, 301, 302, 303, 400, 401, 402, 403, 500, 501,
+func ForEachDetectionElement(detElemIDHandler func(deid int)) {
+	for _, deid := range []int{100, 101, 102, 103, 200, 201, 202, 203, 300, 301, 302, 303, 400, 401, 402, 403, 500, 501,
 		502, 503, 504, 505, 506, 507, 508, 509, 510, 511, 512, 513, 514, 515, 516, 517, 600, 601,
 		602, 603, 604, 605, 606, 607, 608, 609, 610, 611, 612, 613, 614, 615, 616, 617, 700, 701,
 		702, 703, 704, 705, 706, 707, 708, 709, 710, 711, 712, 713, 714, 715, 716, 717, 718, 719,
@@ -46,15 +49,15 @@ func ForEachDetectionElement(detElemIDHandler func(detElemID int)) {
 		904, 905, 906, 907, 908, 909, 910, 911, 912, 913, 914, 915, 916, 917, 918, 919, 920, 921,
 		922, 923, 924, 925, 1000, 1001, 1002, 1003, 1004, 1005, 1006, 1007, 1008, 1009, 1010, 1011, 1012, 1013,
 		1014, 1015, 1016, 1017, 1018, 1019, 1020, 1021, 1022, 1023, 1024, 1025} {
-		detElemIDHandler(detElemID)
+		detElemIDHandler(deid)
 	}
 }
 
 // ForOneDetectionElementOfEachSegmentationType loops over one detection element per segmentation type
 // and call the detElemIdHandler function for each of them
-func ForOneDetectionElementOfEachSegmentationType(detElemIDHandler func(detElemID int)) {
-	for _, detElemID := range []int{100, 300, 500, 501, 502, 503, 504, 600, 601, 602, 700, 701, 702, 703, 704, 705, 706, 902, 903, 904, 905} {
-		detElemIDHandler(detElemID)
+func ForOneDetectionElementOfEachSegmentationType(detElemIDHandler func(deid int)) {
+	for _, deid := range []int{100, 300, 500, 501, 502, 503, 504, 600, 601, 602, 700, 701, 702, 703, 704, 705, 706, 902, 903, 904, 905} {
+		detElemIDHandler(deid)
 	}
 }
 
@@ -113,4 +116,18 @@ func ComputePadBBox(cseg CathodeSegmentation, paduid PadUID) geo.BBox {
 		return nil
 	}
 	return bbox
+}
+
+// NewSegmentation creates a segmentation object for the given detection element plane
+func NewCathodeSegmentation(deid int, isBendingPlane bool) CathodeSegmentation {
+	segType, err := detElemID2SegType(deid)
+	if err != nil {
+		return nil
+	}
+	builder := getCathodeSegmentationBuilder(segType)
+	if builder == nil {
+		return nil
+	}
+	seg := builder.Build(isBendingPlane, deid)
+	return seg
 }
