+++ mapping/factory.go
@@ -5,12 +5,12 @@ import (
 )
 
 type cathodeSegmentationBuilder interface {
-	Build(isBendingPlane bool) CathodeSegmentation
+	Build(isBendingPlane bool, deid int) CathodeSegmentation
 }
 
 var builderRegistry map[int]cathodeSegmentationBuilder
 
-func registerCathodeSegmentationBuilder(segType int, builder cathodeSegmentationBuilder) {
+func RegisterCathodeSegmentationBuilder(segType int, builder cathodeSegmentationBuilder) {
 	if builderRegistry == nil {
 		builderRegistry = make(map[int]cathodeSegmentationBuilder)
 	}
@@ -19,6 +19,7 @@ func registerCathodeSegmentationBuilder(segType int, builder cathodeSegmentation
 		fmt.Printf("already got a build for segType %d : will not override it", segType)
 		return
 	}
+	fmt.Printf("register builder for segType %d ", segType)
 	builderRegistry[segType] = builder
 }
