+++ mapping/impl4/createSegType12.go
@@ -1,81 +1,78 @@
-package mapping
+package impl4
 
-type createSegType17 struct{}
+import "github.com/aphecetche/pigiron/mapping"
 
-func (seg createSegType17) Build(isBendingPlane bool) CathodeSegmentation {
+type createSegType12 struct{}
+
+func (seg createSegType12) Build(isBendingPlane bool, deid int) mapping.CathodeSegmentation {
 	if isBendingPlane {
-		return newCathodeSegmentation(17, true,
+		return newCathodeSegmentation(deid, 12, true,
 			[]padGroup{
-				{1, 3, 0, -120, -20},
-				{2, 12, 0, -117.5, -20},
-				{3, 6, 0, -112.5, -20},
-				{4, 13, 0, -110, -20},
-				{5, 4, 0, -105, -20},
-				{6, 3, 0, -100, -20},
-				{7, 12, 0, -97.5, -20},
-				{8, 6, 0, -92.5, -20},
-				{9, 13, 0, -90, -20},
-				{10, 4, 0, -85, -20},
-				{18, 3, 0, -80, -20},
-				{19, 12, 0, -77.5, -20},
-				{20, 6, 0, -72.5, -20},
-				{21, 13, 0, -70, -20},
-				{22, 4, 0, -65, -20},
-				{23, 3, 0, -60, -20},
-				{24, 12, 0, -57.5, -20},
-				{25, 6, 0, -52.5, -20},
-				{26, 13, 0, -50, -20},
-				{27, 4, 0, -45, -20},
-				{35, 3, 1, -40, -20},
-				{36, 12, 1, -35, -20},
-				{37, 6, 1, -25, -20},
-				{38, 13, 1, -20, -20},
-				{39, 4, 1, -10, -20},
-				{103, 8, 2, 80, -20},
-				{104, 8, 2, 100, -20},
-				{107, 8, 2, 40, -20},
-				{108, 8, 2, 60, -20},
-				{112, 3, 1, 0, -20},
-				{113, 12, 1, 5, -20},
-				{114, 6, 1, 15, -20},
-				{115, 13, 1, 20, -20},
-				{116, 4, 1, 30, -20},
-				{201, 0, 2, 100, -4},
-				{202, 7, 2, 90, 4},
-				{203, 5, 2, 80, -4},
-				{206, 0, 2, 60, -4},
-				{207, 7, 2, 50, 4},
-				{208, 5, 2, 40, -4},
-				{211, 2, 1, 30, 0},
-				{212, 11, 1, 20, 0},
-				{213, 9, 1, 15, 4},
-				{214, 10, 1, 5, 0},
-				{215, 1, 1, 0, 0},
-				{308, 2, 0, -85, 0},
-				{309, 11, 0, -90, 0},
-				{310, 9, 0, -92.5, 4},
-				{311, 10, 0, -97.5, 0},
-				{312, 1, 0, -100, 0},
-				{313, 2, 0, -105, 0},
-				{314, 11, 0, -110, 0},
-				{315, 9, 0, -112.5, 4},
-				{316, 10, 0, -117.5, 0},
-				{317, 1, 0, -120, 0},
-				{325, 2, 0, -45, 0},
-				{326, 11, 0, -50, 0},
-				{327, 9, 0, -52.5, 4},
-				{328, 10, 0, -57.5, 0},
-				{329, 1, 0, -60, 0},
-				{330, 2, 0, -65, 0},
-				{331, 11, 0, -70, 0},
-				{332, 9, 0, -72.5, 4},
-				{333, 10, 0, -77.5, 0},
-				{334, 1, 0, -80, 0},
-				{338, 2, 1, -10, 0},
-				{339, 11, 1, -20, 0},
-				{340, 9, 1, -25, 4},
-				{341, 10, 1, -35, 0},
-				{342, 1, 1, -40, 0},
+				{1, 3, 0, -100, -20},
+				{2, 12, 0, -97.5, -20},
+				{3, 6, 0, -92.5, -20},
+				{4, 13, 0, -90, -20},
+				{5, 4, 0, -85, -20},
+				{6, 3, 0, -80, -20},
+				{7, 12, 0, -77.5, -20},
+				{8, 6, 0, -72.5, -20},
+				{9, 13, 0, -70, -20},
+				{10, 4, 0, -65, -20},
+				{18, 3, 0, -60, -20},
+				{19, 12, 0, -57.5, -20},
+				{20, 6, 0, -52.5, -20},
+				{21, 13, 0, -50, -20},
+				{22, 4, 0, -45, -20},
+				{23, 3, 0, -40, -20},
+				{24, 12, 0, -37.5, -20},
+				{25, 6, 0, -32.5, -20},
+				{26, 13, 0, -30, -20},
+				{27, 4, 0, -25, -20},
+				{103, 8, 2, 60, -20},
+				{104, 8, 2, 80, -20},
+				{108, 3, 1, 20, -20},
+				{109, 12, 1, 25, -20},
+				{110, 6, 1, 35, -20},
+				{111, 13, 1, 40, -20},
+				{112, 4, 1, 50, -20},
+				{116, 3, 1, -20, -20},
+				{117, 12, 1, -15, -20},
+				{118, 6, 1, -5, -20},
+				{119, 13, 1, 0, -20},
+				{120, 4, 1, 10, -20},
+				{201, 0, 2, 80, -4},
+				{202, 7, 2, 70, 4},
+				{203, 5, 2, 60, -4},
+				{206, 2, 1, 50, 0},
+				{207, 11, 1, 40, 0},
+				{208, 9, 1, 35, 4},
+				{209, 10, 1, 25, 0},
+				{210, 1, 1, 20, 0},
+				{215, 2, 1, 10, 0},
+				{216, 11, 1, 0, 0},
+				{217, 9, 1, -5, 4},
+				{218, 10, 1, -15, 0},
+				{219, 1, 1, -20, 0},
+				{308, 2, 0, -65, 0},
+				{309, 11, 0, -70, 0},
+				{310, 9, 0, -72.5, 4},
+				{311, 10, 0, -77.5, 0},
+				{312, 1, 0, -80, 0},
+				{313, 2, 0, -85, 0},
+				{314, 11, 0, -90, 0},
+				{315, 9, 0, -92.5, 4},
+				{316, 10, 0, -97.5, 0},
+				{317, 1, 0, -100, 0},
+				{325, 2, 0, -25, 0},
+				{326, 11, 0, -30, 0},
+				{327, 9, 0, -32.5, 4},
+				{328, 10, 0, -37.5, 0},
+				{329, 1, 0, -40, 0},
+				{330, 2, 0, -45, 0},
+				{331, 11, 0, -50, 0},
+				{332, 9, 0, -52.5, 4},
+				{333, 10, 0, -57.5, 0},
+				{334, 1, 0, -60, 0},
 			},
 			[]padGroupType{
 				/* L10 */ NewPadGroupType(2, 48, []int{15, 16, 14, 17, 13, 18, 12, 19, 11, 20, 10, 21, 9, 22, 8, 23, 7, 24, 6, 25, 5, 26, 4, 27, 3, 28, 2, 29, 1, 30, 0, 31, -1, 48, -1, 49, -1, 50, -1, 51, -1, 52, -1, 53, -1, 54, -1, 55, -1, 56, -1, 57, -1, 58, -1, 59, -1, 60, -1, 61, -1, 62, -1, 63, -1, 32, -1, 33, -1, 34, -1, 35, -1, 36, -1, 37, -1, 38, -1, 39, -1, 40, -1, 41, -1, 42, -1, 43, -1, 44, -1, 45, -1, 46, -1, 47}),
@@ -99,80 +96,69 @@ func (seg createSegType17) Build(isBendingPlane bool) CathodeSegmentation {
 				{10, 0.5},
 			})
 	}
-	return newCathodeSegmentation(17, false,
+	return newCathodeSegmentation(deid, 12, false,
 		[]padGroup{
-			{1035, 5, 0, -85.7142868, -20},
-			{1036, 5, 0, -91.42857361, -20},
-			{1037, 5, 0, -97.14286041, -20},
-			{1038, 5, 0, -102.8571396, -20},
-			{1039, 5, 0, -108.5714264, -20},
-			{1040, 5, 0, -114.2857132, -20},
-			{1041, 5, 0, -120, -20},
-			{1052, 5, 0, -45.7142868, -20},
-			{1053, 5, 0, -51.42856979, -20},
-			{1054, 5, 0, -57.1428566, -20},
-			{1055, 5, 0, -62.8571434, -20},
-			{1056, 5, 0, -68.57142639, -20},
-			{1057, 5, 0, -74.2857132, -20},
-			{1058, 5, 0, -80, -20},
-			{1064, 15, 1, -10, -20},
-			{1065, 19, 1, -20, -20},
-			{1066, 18, 1, -31.4285717, -20},
-			{1067, 14, 1, -40, -20},
-			{1125, 11, 2, 100, -20},
-			{1126, 10, 2, 80, -20},
-			{1129, 11, 2, 60, -20},
-			{1130, 10, 2, 40, -20},
-			{1133, 1, 1, 25.7142849, -20},
-			{1134, 6, 1, 14.28571415, -20},
-			{1135, 0, 1, 4.440892099e-15, -20},
-			{1228, 8, 2, 80, 0},
-			{1229, 9, 2, 100, 0},
-			{1233, 8, 2, 40, 0},
-			{1234, 9, 2, 60, 0},
-			{1240, 16, 1, -7.105427358e-15, -5},
-			{1241, 12, 1, 8.571428299, -5},
-			{1242, 13, 1, 20, -5},
-			{1243, 17, 1, 30, -5},
-			{1325, 4, 0, -120, 0},
-			{1326, 4, 0, -114.2857132, 0},
-			{1327, 4, 0, -108.5714264, 0},
-			{1328, 4, 0, -102.8571396, 0},
-			{1329, 4, 0, -97.14286041, 0},
-			{1330, 4, 0, -91.42857361, 0},
-			{1331, 4, 0, -85.7142868, 0},
-			{1342, 4, 0, -80, 0},
-			{1343, 4, 0, -74.2857132, 0},
-			{1344, 4, 0, -68.57142639, 0},
-			{1345, 4, 0, -62.8571434, 0},
-			{1346, 4, 0, -57.1428566, 0},
-			{1347, 4, 0, -51.42856979, 0},
-			{1348, 4, 0, -45.7142868, 0},
-			{1359, 2, 1, -40, 0},
-			{1360, 7, 1, -25.7142849, 0},
-			{1361, 3, 1, -14.28571415, 0},
+			{1035, 3, 0, -65.7142868, -20},
+			{1036, 3, 0, -71.42857361, -20},
+			{1037, 3, 0, -77.14286041, -20},
+			{1038, 3, 0, -82.85713959, -20},
+			{1039, 3, 0, -88.57142639, -20},
+			{1040, 3, 0, -94.2857132, -20},
+			{1041, 3, 0, -100, -20},
+			{1052, 3, 0, -25.7142849, -20},
+			{1053, 3, 0, -31.4285717, -20},
+			{1054, 3, 0, -37.1428566, -20},
+			{1055, 3, 0, -42.8571434, -20},
+			{1056, 3, 0, -48.57143021, -20},
+			{1057, 3, 0, -54.2857132, -20},
+			{1058, 3, 0, -60, -20},
+			{1125, 8, 2, 80, -20},
+			{1126, 7, 2, 60, -20},
+			{1129, 1, 1, 45.7142868, -20},
+			{1130, 4, 1, 34.2857132, -20},
+			{1131, 0, 1, 20, -20},
+			{1137, 1, 1, 5.714285851, -20},
+			{1138, 4, 1, -5.714285851, -20},
+			{1139, 0, 1, -20, -20},
+			{1228, 5, 2, 60, 0},
+			{1229, 6, 2, 80, 0},
+			{1235, 11, 1, 20, -5},
+			{1236, 9, 1, 28.5714283, -5},
+			{1237, 10, 1, 40, -5},
+			{1238, 12, 1, 50, -5},
+			{1244, 11, 1, -20, -5},
+			{1245, 9, 1, -11.4285717, -5},
+			{1246, 10, 1, -8.000008656e-09, -5},
+			{1247, 12, 1, 10, -5},
+			{1325, 2, 0, -100, 0},
+			{1326, 2, 0, -94.2857132, 0},
+			{1327, 2, 0, -88.57142639, 0},
+			{1328, 2, 0, -82.85713959, 0},
+			{1329, 2, 0, -77.14286041, 0},
+			{1330, 2, 0, -71.42857361, 0},
+			{1331, 2, 0, -65.7142868, 0},
+			{1342, 2, 0, -60, 0},
+			{1343, 2, 0, -54.2857132, 0},
+			{1344, 2, 0, -48.57143021, 0},
+			{1345, 2, 0, -42.8571434, 0},
+			{1346, 2, 0, -37.1428566, 0},
+			{1347, 2, 0, -31.4285717, 0},
+			{1348, 2, 0, -25.7142849, 0},
 		},
 		[]padGroupType{
 			/* L1 */ NewPadGroupType(20, 4, []int{3, 7, 11, 15, 18, 21, 24, 27, 30, 49, 52, 55, 58, 61, 32, 35, 38, 41, 44, 47, 2, 6, 10, 14, 17, 20, 23, 26, 29, 48, 51, 54, 57, 60, 63, 34, 37, 40, 43, 46, 1, 5, 9, 13, 16, 19, 22, 25, 28, 31, 50, 53, 56, 59, 62, 33, 36, 39, 42, 45, 0, 4, 8, 12, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1}),
 			/* L2 */ NewPadGroupType(20, 4, []int{2, 5, 8, 11, 14, 17, 20, 23, 26, 29, 48, 51, 54, 57, 60, 63, 35, 39, 43, 47, 1, 4, 7, 10, 13, 16, 19, 22, 25, 28, 31, 50, 53, 56, 59, 62, 34, 38, 42, 46, 0, 3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 49, 52, 55, 58, 61, 33, 37, 41, 45, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 32, 36, 40, 44}),
-			/* L3 */ NewPadGroupType(20, 4, []int{44, 40, 36, 32, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 45, 41, 37, 33, 61, 58, 55, 52, 49, 30, 27, 24, 21, 18, 15, 12, 9, 6, 3, 0, 46, 42, 38, 34, 62, 59, 56, 53, 50, 31, 28, 25, 22, 19, 16, 13, 10, 7, 4, 1, 47, 43, 39, 35, 63, 60, 57, 54, 51, 48, 29, 26, 23, 20, 17, 14, 11, 8, 5, 2}),
-			/* L4 */ NewPadGroupType(20, 4, []int{-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 12, 8, 4, 0, 45, 42, 39, 36, 33, 62, 59, 56, 53, 50, 31, 28, 25, 22, 19, 16, 13, 9, 5, 1, 46, 43, 40, 37, 34, 63, 60, 57, 54, 51, 48, 29, 26, 23, 20, 17, 14, 10, 6, 2, 47, 44, 41, 38, 35, 32, 61, 58, 55, 52, 49, 30, 27, 24, 21, 18, 15, 11, 7, 3}),
 			/* O1 */ NewPadGroupType(8, 8, []int{40, 32, 56, 48, 24, 16, 8, 0, 41, 33, 57, 49, 25, 17, 9, 1, 42, 34, 58, 50, 26, 18, 10, 2, 43, 35, 59, 51, 27, 19, 11, 3, 44, 36, 60, 52, 28, 20, 12, 4, 45, 37, 61, 53, 29, 21, 13, 5, 46, 38, 62, 54, 30, 22, 14, 6, 47, 39, 63, 55, 31, 23, 15, 7}),
 			/* O2 */ NewPadGroupType(8, 8, []int{7, 15, 23, 31, 55, 63, 39, 47, 6, 14, 22, 30, 54, 62, 38, 46, 5, 13, 21, 29, 53, 61, 37, 45, 4, 12, 20, 28, 52, 60, 36, 44, 3, 11, 19, 27, 51, 59, 35, 43, 2, 10, 18, 26, 50, 58, 34, 42, 1, 9, 17, 25, 49, 57, 33, 41, 0, 8, 16, 24, 48, 56, 32, 40}),
 			/* O3 */ NewPadGroupType(16, 4, []int{3, 7, 11, 15, 19, 23, 27, 31, 51, 55, 59, 63, 35, 39, 43, 47, 2, 6, 10, 14, 18, 22, 26, 30, 50, 54, 58, 62, 34, 38, 42, 46, 1, 5, 9, 13, 17, 21, 25, 29, 49, 53, 57, 61, 33, 37, 41, 45, 0, 4, 8, 12, 16, 20, 24, 28, 48, 52, 56, 60, 32, 36, 40, 44}),
-			/* O4 */ NewPadGroupType(16, 4, []int{44, 40, 36, 32, 60, 56, 52, 48, 28, 24, 20, 16, 12, 8, 4, 0, 45, 41, 37, 33, 61, 57, 53, 49, 29, 25, 21, 17, 13, 9, 5, 1, 46, 42, 38, 34, 62, 58, 54, 50, 30, 26, 22, 18, 14, 10, 6, 2, 47, 43, 39, 35, 63, 59, 55, 51, 31, 27, 23, 19, 15, 11, 7, 3}),
 			/* O5 */ NewPadGroupType(28, 2, []int{47, 45, 43, 41, 39, 37, 35, 33, 63, 61, 59, 57, 55, 53, 51, 49, 31, 29, 27, 25, 23, 21, 19, 17, 15, 13, 11, 9, 46, 44, 42, 40, 38, 36, 34, 32, 62, 60, 58, 56, 54, 52, 50, 48, 30, 28, 26, 24, 22, 20, 18, 16, 14, 12, 10, 8}),
 			/* O6 */ NewPadGroupType(28, 2, []int{39, 37, 35, 33, 63, 61, 59, 57, 55, 53, 51, 49, 31, 29, 27, 25, 23, 21, 19, 17, 15, 13, 11, 9, 7, 5, 3, 1, 38, 36, 34, 32, 62, 60, 58, 56, 54, 52, 50, 48, 30, 28, 26, 24, 22, 20, 18, 16, 14, 12, 10, 8, 6, 4, 2, 0}),
 			/* O7 */ NewPadGroupType(28, 2, []int{0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 48, 50, 52, 54, 56, 58, 60, 62, 32, 34, 36, 38, 1, 3, 5, 7, 9, 11, 13, 15, 17, 19, 21, 23, 25, 27, 29, 31, 49, 51, 53, 55, 57, 59, 61, 63, 33, 35, 37, 39}),
 			/* O8 */ NewPadGroupType(28, 2, []int{8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 48, 50, 52, 54, 56, 58, 60, 62, 32, 34, 36, 38, 40, 42, 44, 46, 9, 11, 13, 15, 17, 19, 21, 23, 25, 27, 29, 31, 49, 51, 53, 55, 57, 59, 61, 63, 33, 35, 37, 39, 41, 43, 45, 47}),
 			/* P1 */ NewPadGroupType(16, 5, []int{47, 46, 41, 36, 63, 58, 53, 48, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 42, 37, 32, 59, 54, 49, 28, 24, 20, 16, 12, 8, 4, 0, -1, -1, 43, 38, 33, 60, 55, 50, 29, 25, 21, 17, 13, 9, 5, 1, -1, -1, 44, 39, 34, 61, 56, 51, 30, 26, 22, 18, 14, 10, 6, 2, -1, -1, 45, 40, 35, 62, 57, 52, 31, 27, 23, 19, 15, 11, 7, 3}),
 			/* P2 */ NewPadGroupType(16, 5, []int{-1, -1, -1, -1, -1, -1, -1, -1, 27, 22, 17, 12, 7, 2, 1, 0, 44, 40, 36, 32, 60, 56, 52, 48, 28, 23, 18, 13, 8, 3, -1, -1, 45, 41, 37, 33, 61, 57, 53, 49, 29, 24, 19, 14, 9, 4, -1, -1, 46, 42, 38, 34, 62, 58, 54, 50, 30, 25, 20, 15, 10, 5, -1, -1, 47, 43, 39, 35, 63, 59, 55, 51, 31, 26, 21, 16, 11, 6, -1, -1}),
-			/* P3 */ NewPadGroupType(14, 5, []int{3, 7, 11, 15, 20, 25, 30, 51, 56, 61, 34, 39, 43, 47, 2, 6, 10, 14, 19, 24, 29, 50, 55, 60, 33, 38, 42, 46, 1, 5, 9, 13, 18, 23, 28, 49, 54, 59, 32, 37, 41, 45, 0, 4, 8, 12, 17, 22, 27, 48, 53, 58, 63, 36, 40, 44, -1, -1, -1, -1, 16, 21, 26, 31, 52, 57, 62, 35, -1, -1}),
-			/* P4 */ NewPadGroupType(14, 5, []int{3, 7, 12, 17, 22, 27, 48, 53, 58, 63, 35, 39, 43, 47, 2, 6, 11, 16, 21, 26, 31, 52, 57, 62, 34, 38, 42, 46, 1, 5, 10, 15, 20, 25, 30, 51, 56, 61, 33, 37, 41, 45, 0, 4, 9, 14, 19, 24, 29, 50, 55, 60, 32, 36, 40, 44, -1, -1, 8, 13, 18, 23, 28, 49, 54, 59, -1, -1, -1, -1}),
 			/* Q1 */ NewPadGroupType(14, 5, []int{-1, -1, -1, -1, 59, 54, 49, 28, 23, 18, 13, 8, -1, -1, 44, 40, 36, 32, 60, 55, 50, 29, 24, 19, 14, 9, 4, 0, 45, 41, 37, 33, 61, 56, 51, 30, 25, 20, 15, 10, 5, 1, 46, 42, 38, 34, 62, 57, 52, 31, 26, 21, 16, 11, 6, 2, 47, 43, 39, 35, 63, 58, 53, 48, 27, 22, 17, 12, 7, 3}),
 			/* Q2 */ NewPadGroupType(14, 5, []int{-1, -1, 35, 62, 57, 52, 31, 26, 21, 16, -1, -1, -1, -1, 44, 40, 36, 63, 58, 53, 48, 27, 22, 17, 12, 8, 4, 0, 45, 41, 37, 32, 59, 54, 49, 28, 23, 18, 13, 9, 5, 1, 46, 42, 38, 33, 60, 55, 50, 29, 24, 19, 14, 10, 6, 2, 47, 43, 39, 34, 61, 56, 51, 30, 25, 20, 15, 11, 7, 3}),
-			/* Q3 */ NewPadGroupType(16, 5, []int{-1, -1, 6, 11, 16, 21, 26, 31, 51, 55, 59, 63, 35, 39, 43, 47, -1, -1, 5, 10, 15, 20, 25, 30, 50, 54, 58, 62, 34, 38, 42, 46, -1, -1, 4, 9, 14, 19, 24, 29, 49, 53, 57, 61, 33, 37, 41, 45, -1, -1, 3, 8, 13, 18, 23, 28, 48, 52, 56, 60, 32, 36, 40, 44, 0, 1, 2, 7, 12, 17, 22, 27, -1, -1, -1, -1, -1, -1, -1, -1}),
-			/* Q4 */ NewPadGroupType(16, 5, []int{3, 7, 11, 15, 19, 23, 27, 31, 52, 57, 62, 35, 40, 45, -1, -1, 2, 6, 10, 14, 18, 22, 26, 30, 51, 56, 61, 34, 39, 44, -1, -1, 1, 5, 9, 13, 17, 21, 25, 29, 50, 55, 60, 33, 38, 43, -1, -1, 0, 4, 8, 12, 16, 20, 24, 28, 49, 54, 59, 32, 37, 42, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 48, 53, 58, 63, 36, 41, 46, 47}),
 		},
 		[]padSize{
 			{0.714285714, 2.5},
@@ -182,5 +168,5 @@ func (seg createSegType17) Build(isBendingPlane bool) CathodeSegmentation {
 }
 
 func init() {
-	registerCathodeSegmentationBuilder(17, createSegType17{})
+	mapping.RegisterCathodeSegmentationBuilder(12, createSegType12{})
 }
