+++ mapping/impl4/createSegType13.go
@@ -1,10 +1,12 @@
-package mapping
+package impl4
 
-type createSegType19 struct{}
+import "github.com/aphecetche/pigiron/mapping"
 
-func (seg createSegType19) Build(isBendingPlane bool) CathodeSegmentation {
+type createSegType13 struct{}
+
+func (seg createSegType13) Build(isBendingPlane bool, deid int) mapping.CathodeSegmentation {
 	if isBendingPlane {
-		return newCathodeSegmentation(19, true,
+		return newCathodeSegmentation(deid, 13, true,
 			[]padGroup{
 				{1, 3, 0, -100, -20},
 				{2, 12, 0, -95, -20},
@@ -20,17 +22,22 @@ func (seg createSegType19) Build(isBendingPlane bool) CathodeSegmentation {
 				{104, 8, 1, 80, -20},
 				{107, 8, 1, 20, -20},
 				{108, 8, 1, 40, -20},
-				{111, 8, 1, -20, -20},
-				{112, 8, 1, 0, -20},
+				{112, 3, 0, -20, -20},
+				{113, 12, 0, -15, -20},
+				{114, 6, 0, -5, -20},
+				{115, 13, 0, 0, -20},
+				{116, 4, 0, 10, -20},
 				{201, 0, 1, 80, -4},
 				{202, 7, 1, 70, 4},
 				{203, 5, 1, 60, -4},
 				{206, 0, 1, 40, -4},
 				{207, 7, 1, 30, 4},
 				{208, 5, 1, 20, -4},
-				{211, 0, 1, 0, -4},
-				{212, 7, 1, -10, 4},
-				{213, 5, 1, -20, -4},
+				{211, 2, 0, 10, 0},
+				{212, 11, 0, 0, 0},
+				{213, 9, 0, -5, 4},
+				{214, 10, 0, -15, 0},
+				{215, 1, 0, -20, 0},
 				{304, 2, 0, -70, 0},
 				{305, 11, 0, -80, 0},
 				{306, 9, 0, -85, 4},
@@ -63,45 +70,55 @@ func (seg createSegType19) Build(isBendingPlane bool) CathodeSegmentation {
 				{10, 0.5},
 			})
 	}
-	return newCathodeSegmentation(19, false,
+	return newCathodeSegmentation(deid, 13, false,
 		[]padGroup{
-			{1030, 8, 0, -70, -20},
-			{1031, 10, 0, -80, -20},
-			{1032, 9, 0, -91.42857361, -20},
-			{1033, 7, 0, -100, -20},
-			{1039, 8, 0, -30, -20},
-			{1040, 10, 0, -40, -20},
-			{1041, 9, 0, -51.42856979, -20},
-			{1042, 7, 0, -60, -20},
-			{1125, 6, 1, 80, -20},
-			{1126, 5, 1, 60, -20},
-			{1129, 6, 1, 40, -20},
-			{1130, 5, 1, 20, -20},
-			{1133, 6, 1, -8.000000662e-09, -20},
-			{1134, 5, 1, -20, -20},
-			{1228, 3, 1, 60, 0},
-			{1229, 4, 1, 80, 0},
-			{1233, 3, 1, 20, 0},
-			{1234, 4, 1, 40, 0},
-			{1238, 3, 1, -20, 0},
-			{1239, 4, 1, -8.000000662e-09, 0},
-			{1325, 0, 0, -100, 0},
-			{1326, 2, 0, -85.7142868, 0},
-			{1327, 1, 0, -74.2857132, 0},
-			{1333, 0, 0, -60, 0},
-			{1334, 2, 0, -45.7142868, 0},
-			{1335, 1, 0, -34.2857132, 0},
+			{1030, 13, 0, -70, -20},
+			{1031, 17, 0, -80, -20},
+			{1032, 16, 0, -91.42857361, -20},
+			{1033, 12, 0, -100, -20},
+			{1039, 13, 0, -30, -20},
+			{1040, 17, 0, -40, -20},
+			{1041, 16, 0, -51.42856979, -20},
+			{1042, 12, 0, -60, -20},
+			{1125, 9, 1, 80, -20},
+			{1126, 8, 1, 60, -20},
+			{1129, 9, 1, 40, -20},
+			{1130, 8, 1, 20, -20},
+			{1133, 1, 0, 5.714285851, -20},
+			{1134, 4, 0, -5.714285851, -20},
+			{1135, 0, 0, -20, -20},
+			{1228, 6, 1, 60, 0},
+			{1229, 7, 1, 80, 0},
+			{1233, 6, 1, 20, 0},
+			{1234, 7, 1, 40, 0},
+			{1240, 14, 0, -20, -5},
+			{1241, 10, 0, -11.4285717, -5},
+			{1242, 11, 0, -8.000008656e-09, -5},
+			{1243, 15, 0, 10, -5},
+			{1325, 2, 0, -100, 0},
+			{1326, 5, 0, -85.7142868, 0},
+			{1327, 3, 0, -74.2857132, 0},
+			{1333, 2, 0, -60, 0},
+			{1334, 5, 0, -45.7142868, 0},
+			{1335, 3, 0, -34.2857132, 0},
 		},
 		[]padGroupType{
+			/* L1 */ NewPadGroupType(20, 4, []int{3, 7, 11, 15, 18, 21, 24, 27, 30, 49, 52, 55, 58, 61, 32, 35, 38, 41, 44, 47, 2, 6, 10, 14, 17, 20, 23, 26, 29, 48, 51, 54, 57, 60, 63, 34, 37, 40, 43, 46, 1, 5, 9, 13, 16, 19, 22, 25, 28, 31, 50, 53, 56, 59, 62, 33, 36, 39, 42, 45, 0, 4, 8, 12, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1}),
+			/* L2 */ NewPadGroupType(20, 4, []int{2, 5, 8, 11, 14, 17, 20, 23, 26, 29, 48, 51, 54, 57, 60, 63, 35, 39, 43, 47, 1, 4, 7, 10, 13, 16, 19, 22, 25, 28, 31, 50, 53, 56, 59, 62, 34, 38, 42, 46, 0, 3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 49, 52, 55, 58, 61, 33, 37, 41, 45, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 32, 36, 40, 44}),
 			/* L3 */ NewPadGroupType(20, 4, []int{44, 40, 36, 32, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 45, 41, 37, 33, 61, 58, 55, 52, 49, 30, 27, 24, 21, 18, 15, 12, 9, 6, 3, 0, 46, 42, 38, 34, 62, 59, 56, 53, 50, 31, 28, 25, 22, 19, 16, 13, 10, 7, 4, 1, 47, 43, 39, 35, 63, 60, 57, 54, 51, 48, 29, 26, 23, 20, 17, 14, 11, 8, 5, 2}),
 			/* L4 */ NewPadGroupType(20, 4, []int{-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 12, 8, 4, 0, 45, 42, 39, 36, 33, 62, 59, 56, 53, 50, 31, 28, 25, 22, 19, 16, 13, 9, 5, 1, 46, 43, 40, 37, 34, 63, 60, 57, 54, 51, 48, 29, 26, 23, 20, 17, 14, 10, 6, 2, 47, 44, 41, 38, 35, 32, 61, 58, 55, 52, 49, 30, 27, 24, 21, 18, 15, 11, 7, 3}),
+			/* O3 */ NewPadGroupType(16, 4, []int{3, 7, 11, 15, 19, 23, 27, 31, 51, 55, 59, 63, 35, 39, 43, 47, 2, 6, 10, 14, 18, 22, 26, 30, 50, 54, 58, 62, 34, 38, 42, 46, 1, 5, 9, 13, 17, 21, 25, 29, 49, 53, 57, 61, 33, 37, 41, 45, 0, 4, 8, 12, 16, 20, 24, 28, 48, 52, 56, 60, 32, 36, 40, 44}),
 			/* O4 */ NewPadGroupType(16, 4, []int{44, 40, 36, 32, 60, 56, 52, 48, 28, 24, 20, 16, 12, 8, 4, 0, 45, 41, 37, 33, 61, 57, 53, 49, 29, 25, 21, 17, 13, 9, 5, 1, 46, 42, 38, 34, 62, 58, 54, 50, 30, 26, 22, 18, 14, 10, 6, 2, 47, 43, 39, 35, 63, 59, 55, 51, 31, 27, 23, 19, 15, 11, 7, 3}),
 			/* O5 */ NewPadGroupType(28, 2, []int{47, 45, 43, 41, 39, 37, 35, 33, 63, 61, 59, 57, 55, 53, 51, 49, 31, 29, 27, 25, 23, 21, 19, 17, 15, 13, 11, 9, 46, 44, 42, 40, 38, 36, 34, 32, 62, 60, 58, 56, 54, 52, 50, 48, 30, 28, 26, 24, 22, 20, 18, 16, 14, 12, 10, 8}),
 			/* O6 */ NewPadGroupType(28, 2, []int{39, 37, 35, 33, 63, 61, 59, 57, 55, 53, 51, 49, 31, 29, 27, 25, 23, 21, 19, 17, 15, 13, 11, 9, 7, 5, 3, 1, 38, 36, 34, 32, 62, 60, 58, 56, 54, 52, 50, 48, 30, 28, 26, 24, 22, 20, 18, 16, 14, 12, 10, 8, 6, 4, 2, 0}),
 			/* O7 */ NewPadGroupType(28, 2, []int{0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 48, 50, 52, 54, 56, 58, 60, 62, 32, 34, 36, 38, 1, 3, 5, 7, 9, 11, 13, 15, 17, 19, 21, 23, 25, 27, 29, 31, 49, 51, 53, 55, 57, 59, 61, 63, 33, 35, 37, 39}),
 			/* O8 */ NewPadGroupType(28, 2, []int{8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 48, 50, 52, 54, 56, 58, 60, 62, 32, 34, 36, 38, 40, 42, 44, 46, 9, 11, 13, 15, 17, 19, 21, 23, 25, 27, 29, 31, 49, 51, 53, 55, 57, 59, 61, 63, 33, 35, 37, 39, 41, 43, 45, 47}),
+			/* P1 */ NewPadGroupType(16, 5, []int{47, 46, 41, 36, 63, 58, 53, 48, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 42, 37, 32, 59, 54, 49, 28, 24, 20, 16, 12, 8, 4, 0, -1, -1, 43, 38, 33, 60, 55, 50, 29, 25, 21, 17, 13, 9, 5, 1, -1, -1, 44, 39, 34, 61, 56, 51, 30, 26, 22, 18, 14, 10, 6, 2, -1, -1, 45, 40, 35, 62, 57, 52, 31, 27, 23, 19, 15, 11, 7, 3}),
+			/* P2 */ NewPadGroupType(16, 5, []int{-1, -1, -1, -1, -1, -1, -1, -1, 27, 22, 17, 12, 7, 2, 1, 0, 44, 40, 36, 32, 60, 56, 52, 48, 28, 23, 18, 13, 8, 3, -1, -1, 45, 41, 37, 33, 61, 57, 53, 49, 29, 24, 19, 14, 9, 4, -1, -1, 46, 42, 38, 34, 62, 58, 54, 50, 30, 25, 20, 15, 10, 5, -1, -1, 47, 43, 39, 35, 63, 59, 55, 51, 31, 26, 21, 16, 11, 6, -1, -1}),
 			/* P3 */ NewPadGroupType(14, 5, []int{3, 7, 11, 15, 20, 25, 30, 51, 56, 61, 34, 39, 43, 47, 2, 6, 10, 14, 19, 24, 29, 50, 55, 60, 33, 38, 42, 46, 1, 5, 9, 13, 18, 23, 28, 49, 54, 59, 32, 37, 41, 45, 0, 4, 8, 12, 17, 22, 27, 48, 53, 58, 63, 36, 40, 44, -1, -1, -1, -1, 16, 21, 26, 31, 52, 57, 62, 35, -1, -1}),
 			/* P4 */ NewPadGroupType(14, 5, []int{3, 7, 12, 17, 22, 27, 48, 53, 58, 63, 35, 39, 43, 47, 2, 6, 11, 16, 21, 26, 31, 52, 57, 62, 34, 38, 42, 46, 1, 5, 10, 15, 20, 25, 30, 51, 56, 61, 33, 37, 41, 45, 0, 4, 9, 14, 19, 24, 29, 50, 55, 60, 32, 36, 40, 44, -1, -1, 8, 13, 18, 23, 28, 49, 54, 59, -1, -1, -1, -1}),
+			/* Q1 */ NewPadGroupType(14, 5, []int{-1, -1, -1, -1, 59, 54, 49, 28, 23, 18, 13, 8, -1, -1, 44, 40, 36, 32, 60, 55, 50, 29, 24, 19, 14, 9, 4, 0, 45, 41, 37, 33, 61, 56, 51, 30, 25, 20, 15, 10, 5, 1, 46, 42, 38, 34, 62, 57, 52, 31, 26, 21, 16, 11, 6, 2, 47, 43, 39, 35, 63, 58, 53, 48, 27, 22, 17, 12, 7, 3}),
+			/* Q2 */ NewPadGroupType(14, 5, []int{-1, -1, 35, 62, 57, 52, 31, 26, 21, 16, -1, -1, -1, -1, 44, 40, 36, 63, 58, 53, 48, 27, 22, 17, 12, 8, 4, 0, 45, 41, 37, 32, 59, 54, 49, 28, 23, 18, 13, 9, 5, 1, 46, 42, 38, 33, 60, 55, 50, 29, 24, 19, 14, 10, 6, 2, 47, 43, 39, 34, 61, 56, 51, 30, 25, 20, 15, 11, 7, 3}),
 			/* Q3 */ NewPadGroupType(16, 5, []int{-1, -1, 6, 11, 16, 21, 26, 31, 51, 55, 59, 63, 35, 39, 43, 47, -1, -1, 5, 10, 15, 20, 25, 30, 50, 54, 58, 62, 34, 38, 42, 46, -1, -1, 4, 9, 14, 19, 24, 29, 49, 53, 57, 61, 33, 37, 41, 45, -1, -1, 3, 8, 13, 18, 23, 28, 48, 52, 56, 60, 32, 36, 40, 44, 0, 1, 2, 7, 12, 17, 22, 27, -1, -1, -1, -1, -1, -1, -1, -1}),
 			/* Q4 */ NewPadGroupType(16, 5, []int{3, 7, 11, 15, 19, 23, 27, 31, 52, 57, 62, 35, 40, 45, -1, -1, 2, 6, 10, 14, 18, 22, 26, 30, 51, 56, 61, 34, 39, 44, -1, -1, 1, 5, 9, 13, 17, 21, 25, 29, 50, 55, 60, 33, 38, 43, -1, -1, 0, 4, 8, 12, 16, 20, 24, 28, 49, 54, 59, 32, 37, 42, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 48, 53, 58, 63, 36, 41, 46, 47}),
 		},
@@ -112,5 +129,5 @@ func (seg createSegType19) Build(isBendingPlane bool) CathodeSegmentation {
 }
 
 func init() {
-	registerCathodeSegmentationBuilder(19, createSegType19{})
+	mapping.RegisterCathodeSegmentationBuilder(13, createSegType13{})
 }
