+++ mapping/impl4/createSegType17.go
@@ -1,61 +1,83 @@
-package mapping
+package impl4
 
-type createSegType10 struct{}
+import "github.com/aphecetche/pigiron/mapping"
 
-func (seg createSegType10) Build(isBendingPlane bool) CathodeSegmentation {
+type createSegType17 struct{}
+
+func (seg createSegType17) Build(isBendingPlane bool, deid int) mapping.CathodeSegmentation {
 	if isBendingPlane {
-		return newCathodeSegmentation(10, true,
+		return newCathodeSegmentation(deid, 17, true,
 			[]padGroup{
-				{1, 3, 0, -100, -20},
-				{2, 12, 0, -97.5, -20},
-				{3, 6, 0, -92.5, -20},
-				{4, 13, 0, -90, -20},
-				{5, 4, 0, -85, -20},
-				{6, 3, 0, -80, -20},
-				{7, 12, 0, -77.5, -20},
-				{8, 6, 0, -72.5, -20},
-				{9, 13, 0, -70, -20},
-				{10, 4, 0, -65, -20},
-				{18, 3, 1, -60, -20},
-				{19, 12, 1, -55, -20},
-				{20, 6, 1, -45, -20},
-				{21, 13, 1, -40, -20},
-				{22, 4, 1, -30, -20},
-				{103, 8, 2, 60, -20},
-				{104, 8, 2, 80, -20},
-				{107, 8, 2, 20, -20},
-				{108, 8, 2, 40, -20},
-				{112, 3, 1, -20, -20},
-				{113, 12, 1, -15, -20},
-				{114, 6, 1, -5, -20},
-				{115, 13, 1, 0, -20},
-				{116, 4, 1, 10, -20},
-				{201, 0, 2, 80, -4},
-				{202, 7, 2, 70, 4},
-				{203, 5, 2, 60, -4},
-				{206, 0, 2, 40, -4},
-				{207, 7, 2, 30, 4},
-				{208, 5, 2, 20, -4},
-				{211, 2, 1, 10, 0},
-				{212, 11, 1, 0, 0},
-				{213, 9, 1, -5, 4},
-				{214, 10, 1, -15, 0},
-				{215, 1, 1, -20, 0},
-				{308, 2, 0, -65, 0},
-				{309, 11, 0, -70, 0},
-				{310, 9, 0, -72.5, 4},
-				{311, 10, 0, -77.5, 0},
-				{312, 1, 0, -80, 0},
-				{313, 2, 0, -85, 0},
-				{314, 11, 0, -90, 0},
-				{315, 9, 0, -92.5, 4},
-				{316, 10, 0, -97.5, 0},
-				{317, 1, 0, -100, 0},
-				{321, 2, 1, -30, 0},
-				{322, 11, 1, -40, 0},
-				{323, 9, 1, -45, 4},
-				{324, 10, 1, -55, 0},
-				{325, 1, 1, -60, 0},
+				{1, 3, 0, -120, -20},
+				{2, 12, 0, -117.5, -20},
+				{3, 6, 0, -112.5, -20},
+				{4, 13, 0, -110, -20},
+				{5, 4, 0, -105, -20},
+				{6, 3, 0, -100, -20},
+				{7, 12, 0, -97.5, -20},
+				{8, 6, 0, -92.5, -20},
+				{9, 13, 0, -90, -20},
+				{10, 4, 0, -85, -20},
+				{18, 3, 0, -80, -20},
+				{19, 12, 0, -77.5, -20},
+				{20, 6, 0, -72.5, -20},
+				{21, 13, 0, -70, -20},
+				{22, 4, 0, -65, -20},
+				{23, 3, 0, -60, -20},
+				{24, 12, 0, -57.5, -20},
+				{25, 6, 0, -52.5, -20},
+				{26, 13, 0, -50, -20},
+				{27, 4, 0, -45, -20},
+				{35, 3, 1, -40, -20},
+				{36, 12, 1, -35, -20},
+				{37, 6, 1, -25, -20},
+				{38, 13, 1, -20, -20},
+				{39, 4, 1, -10, -20},
+				{103, 8, 2, 80, -20},
+				{104, 8, 2, 100, -20},
+				{107, 8, 2, 40, -20},
+				{108, 8, 2, 60, -20},
+				{112, 3, 1, 0, -20},
+				{113, 12, 1, 5, -20},
+				{114, 6, 1, 15, -20},
+				{115, 13, 1, 20, -20},
+				{116, 4, 1, 30, -20},
+				{201, 0, 2, 100, -4},
+				{202, 7, 2, 90, 4},
+				{203, 5, 2, 80, -4},
+				{206, 0, 2, 60, -4},
+				{207, 7, 2, 50, 4},
+				{208, 5, 2, 40, -4},
+				{211, 2, 1, 30, 0},
+				{212, 11, 1, 20, 0},
+				{213, 9, 1, 15, 4},
+				{214, 10, 1, 5, 0},
+				{215, 1, 1, 0, 0},
+				{308, 2, 0, -85, 0},
+				{309, 11, 0, -90, 0},
+				{310, 9, 0, -92.5, 4},
+				{311, 10, 0, -97.5, 0},
+				{312, 1, 0, -100, 0},
+				{313, 2, 0, -105, 0},
+				{314, 11, 0, -110, 0},
+				{315, 9, 0, -112.5, 4},
+				{316, 10, 0, -117.5, 0},
+				{317, 1, 0, -120, 0},
+				{325, 2, 0, -45, 0},
+				{326, 11, 0, -50, 0},
+				{327, 9, 0, -52.5, 4},
+				{328, 10, 0, -57.5, 0},
+				{329, 1, 0, -60, 0},
+				{330, 2, 0, -65, 0},
+				{331, 11, 0, -70, 0},
+				{332, 9, 0, -72.5, 4},
+				{333, 10, 0, -77.5, 0},
+				{334, 1, 0, -80, 0},
+				{338, 2, 1, -10, 0},
+				{339, 11, 1, -20, 0},
+				{340, 9, 1, -25, 4},
+				{341, 10, 1, -35, 0},
+				{342, 1, 1, -40, 0},
 			},
 			[]padGroupType{
 				/* L10 */ NewPadGroupType(2, 48, []int{15, 16, 14, 17, 13, 18, 12, 19, 11, 20, 10, 21, 9, 22, 8, 23, 7, 24, 6, 25, 5, 26, 4, 27, 3, 28, 2, 29, 1, 30, 0, 31, -1, 48, -1, 49, -1, 50, -1, 51, -1, 52, -1, 53, -1, 54, -1, 55, -1, 56, -1, 57, -1, 58, -1, 59, -1, 60, -1, 61, -1, 62, -1, 63, -1, 32, -1, 33, -1, 34, -1, 35, -1, 36, -1, 37, -1, 38, -1, 39, -1, 40, -1, 41, -1, 42, -1, 43, -1, 44, -1, 45, -1, 46, -1, 47}),
@@ -79,44 +101,58 @@ func (seg createSegType10) Build(isBendingPlane bool) CathodeSegmentation {
 				{10, 0.5},
 			})
 	}
-	return newCathodeSegmentation(10, false,
+	return newCathodeSegmentation(deid, 17, false,
 		[]padGroup{
-			{1035, 5, 0, -65.7142868, -20},
-			{1036, 5, 0, -71.42857361, -20},
-			{1037, 5, 0, -77.14286041, -20},
-			{1038, 5, 0, -82.85713959, -20},
-			{1039, 5, 0, -88.57142639, -20},
-			{1040, 5, 0, -94.2857132, -20},
-			{1041, 5, 0, -100, -20},
-			{1047, 15, 1, -30, -20},
-			{1048, 19, 1, -40, -20},
-			{1049, 18, 1, -51.42856979, -20},
-			{1050, 14, 1, -60, -20},
-			{1125, 11, 2, 80, -20},
-			{1126, 10, 2, 60, -20},
-			{1129, 11, 2, 40, -20},
-			{1130, 10, 2, 20, -20},
-			{1133, 1, 1, 5.714285851, -20},
-			{1134, 6, 1, -5.714285851, -20},
-			{1135, 0, 1, -20, -20},
-			{1228, 8, 2, 60, 0},
-			{1229, 9, 2, 80, 0},
-			{1233, 8, 2, 20, 0},
-			{1234, 9, 2, 40, 0},
-			{1240, 16, 1, -20, -5},
-			{1241, 12, 1, -11.4285717, -5},
-			{1242, 13, 1, -8.000008656e-09, -5},
-			{1243, 17, 1, 10, -5},
-			{1325, 4, 0, -100, 0},
-			{1326, 4, 0, -94.2857132, 0},
-			{1327, 4, 0, -88.57142639, 0},
-			{1328, 4, 0, -82.85713959, 0},
-			{1329, 4, 0, -77.14286041, 0},
-			{1330, 4, 0, -71.42857361, 0},
-			{1331, 4, 0, -65.7142868, 0},
-			{1342, 2, 1, -60, 0},
-			{1343, 7, 1, -45.7142868, 0},
-			{1344, 3, 1, -34.2857132, 0},
+			{1035, 5, 0, -85.7142868, -20},
+			{1036, 5, 0, -91.42857361, -20},
+			{1037, 5, 0, -97.14286041, -20},
+			{1038, 5, 0, -102.8571396, -20},
+			{1039, 5, 0, -108.5714264, -20},
+			{1040, 5, 0, -114.2857132, -20},
+			{1041, 5, 0, -120, -20},
+			{1052, 5, 0, -45.7142868, -20},
+			{1053, 5, 0, -51.42856979, -20},
+			{1054, 5, 0, -57.1428566, -20},
+			{1055, 5, 0, -62.8571434, -20},
+			{1056, 5, 0, -68.57142639, -20},
+			{1057, 5, 0, -74.2857132, -20},
+			{1058, 5, 0, -80, -20},
+			{1064, 15, 1, -10, -20},
+			{1065, 19, 1, -20, -20},
+			{1066, 18, 1, -31.4285717, -20},
+			{1067, 14, 1, -40, -20},
+			{1125, 11, 2, 100, -20},
+			{1126, 10, 2, 80, -20},
+			{1129, 11, 2, 60, -20},
+			{1130, 10, 2, 40, -20},
+			{1133, 1, 1, 25.7142849, -20},
+			{1134, 6, 1, 14.28571415, -20},
+			{1135, 0, 1, 4.440892099e-15, -20},
+			{1228, 8, 2, 80, 0},
+			{1229, 9, 2, 100, 0},
+			{1233, 8, 2, 40, 0},
+			{1234, 9, 2, 60, 0},
+			{1240, 16, 1, -7.105427358e-15, -5},
+			{1241, 12, 1, 8.571428299, -5},
+			{1242, 13, 1, 20, -5},
+			{1243, 17, 1, 30, -5},
+			{1325, 4, 0, -120, 0},
+			{1326, 4, 0, -114.2857132, 0},
+			{1327, 4, 0, -108.5714264, 0},
+			{1328, 4, 0, -102.8571396, 0},
+			{1329, 4, 0, -97.14286041, 0},
+			{1330, 4, 0, -91.42857361, 0},
+			{1331, 4, 0, -85.7142868, 0},
+			{1342, 4, 0, -80, 0},
+			{1343, 4, 0, -74.2857132, 0},
+			{1344, 4, 0, -68.57142639, 0},
+			{1345, 4, 0, -62.8571434, 0},
+			{1346, 4, 0, -57.1428566, 0},
+			{1347, 4, 0, -51.42856979, 0},
+			{1348, 4, 0, -45.7142868, 0},
+			{1359, 2, 1, -40, 0},
+			{1360, 7, 1, -25.7142849, 0},
+			{1361, 3, 1, -14.28571415, 0},
 		},
 		[]padGroupType{
 			/* L1 */ NewPadGroupType(20, 4, []int{3, 7, 11, 15, 18, 21, 24, 27, 30, 49, 52, 55, 58, 61, 32, 35, 38, 41, 44, 47, 2, 6, 10, 14, 17, 20, 23, 26, 29, 48, 51, 54, 57, 60, 63, 34, 37, 40, 43, 46, 1, 5, 9, 13, 16, 19, 22, 25, 28, 31, 50, 53, 56, 59, 62, 33, 36, 39, 42, 45, 0, 4, 8, 12, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1}),
@@ -148,5 +184,5 @@ func (seg createSegType10) Build(isBendingPlane bool) CathodeSegmentation {
 }
 
 func init() {
-	registerCathodeSegmentationBuilder(10, createSegType10{})
+	mapping.RegisterCathodeSegmentationBuilder(17, createSegType17{})
 }
