+++ mapping/impl4/createSegType5.go
@@ -1,10 +1,12 @@
-package mapping
+package impl4
+
+import "github.com/aphecetche/pigiron/mapping"
 
 type createSegType5 struct{}
 
-func (seg createSegType5) Build(isBendingPlane bool) CathodeSegmentation {
+func (seg createSegType5) Build(isBendingPlane bool, deid int) mapping.CathodeSegmentation {
 	if isBendingPlane {
-		return newCathodeSegmentation(5, true,
+		return newCathodeSegmentation(deid, 5, true,
 			[]padGroup{
 				{4, 2, 0, 20, -20},
 				{5, 8, 0, 25, -20},
@@ -53,7 +55,7 @@ func (seg createSegType5) Build(isBendingPlane bool) CathodeSegmentation {
 				{5, 0.5},
 			})
 	}
-	return newCathodeSegmentation(5, false,
+	return newCathodeSegmentation(deid, 5, false,
 		[]padGroup{
 			{1025, 1, 0, 45.7142868, -20},
 			{1026, 2, 0, 34.2857132, -20},
@@ -92,5 +94,5 @@ func (seg createSegType5) Build(isBendingPlane bool) CathodeSegmentation {
 }
 
 func init() {
-	registerCathodeSegmentationBuilder(5, createSegType5{})
+	mapping.RegisterCathodeSegmentationBuilder(5, createSegType5{})
 }
