+++ mapping/impl4/createSegType6.go
@@ -1,61 +1,33 @@
-package mapping
+package impl4
 
-type createSegType9 struct{}
+import "github.com/aphecetche/pigiron/mapping"
 
-func (seg createSegType9) Build(isBendingPlane bool) CathodeSegmentation {
+type createSegType6 struct{}
+
+func (seg createSegType6) Build(isBendingPlane bool, deid int) mapping.CathodeSegmentation {
 	if isBendingPlane {
-		return newCathodeSegmentation(9, true,
+		return newCathodeSegmentation(deid, 6, true,
 			[]padGroup{
-				{1, 2, 0, -80, -20},
-				{2, 8, 0, -77.5, -20},
-				{3, 4, 0, -72.5, -20},
-				{4, 9, 0, -70, -20},
-				{5, 3, 0, -65, -20},
-				{6, 2, 0, -60, -20},
-				{7, 8, 0, -57.5, -20},
-				{8, 4, 0, -52.5, -20},
-				{9, 9, 0, -50, -20},
-				{10, 3, 0, -45, -20},
-				{104, 2, 1, 40, -20},
-				{105, 8, 1, 45, -20},
-				{106, 4, 1, 55, -20},
-				{107, 9, 1, 60, -20},
-				{108, 3, 1, 70, -20},
-				{112, 2, 1, 0, -20},
-				{113, 8, 1, 5, -20},
-				{114, 4, 1, 15, -20},
-				{115, 9, 1, 20, -20},
-				{116, 3, 1, 30, -20},
-				{120, 2, 1, -40, -20},
-				{121, 8, 1, -35, -20},
-				{122, 4, 1, -25, -20},
-				{123, 9, 1, -20, -20},
-				{124, 3, 1, -10, -20},
-				{201, 1, 1, 70, 0},
-				{202, 7, 1, 60, 0},
-				{203, 5, 1, 55, 4},
-				{204, 6, 1, 45, 0},
-				{205, 0, 1, 40, 0},
-				{210, 1, 1, 30, 0},
-				{211, 7, 1, 20, 0},
-				{212, 5, 1, 15, 4},
-				{213, 6, 1, 5, 0},
-				{214, 0, 1, 0, 0},
-				{219, 1, 1, -10, 0},
-				{220, 7, 1, -20, 0},
-				{221, 5, 1, -25, 4},
-				{222, 6, 1, -35, 0},
-				{223, 0, 1, -40, 0},
-				{308, 1, 0, -45, 0},
-				{309, 7, 0, -50, 0},
-				{310, 5, 0, -52.5, 4},
-				{311, 6, 0, -57.5, 0},
-				{312, 0, 0, -60, 0},
-				{313, 1, 0, -65, 0},
-				{314, 7, 0, -70, 0},
-				{315, 5, 0, -72.5, 4},
-				{316, 6, 0, -77.5, 0},
-				{317, 0, 0, -80, 0},
+				{4, 2, 0, 0, -20},
+				{5, 8, 0, 5, -20},
+				{6, 4, 0, 15, -20},
+				{7, 9, 0, 20, -20},
+				{8, 3, 0, 30, -20},
+				{12, 2, 0, -40, -20},
+				{13, 8, 0, -35, -20},
+				{14, 4, 0, -25, -20},
+				{15, 9, 0, -20, -20},
+				{16, 3, 0, -10, -20},
+				{101, 1, 0, 30, 0},
+				{102, 7, 0, 20, 0},
+				{103, 5, 0, 15, 4},
+				{104, 6, 0, 5, 0},
+				{105, 0, 0, 0, 0},
+				{110, 1, 0, -10, 0},
+				{111, 7, 0, -20, 0},
+				{112, 5, 0, -25, 4},
+				{113, 6, 0, -35, 0},
+				{114, 0, 0, -40, 0},
 			},
 			[]padGroupType{
 				/* L5 */ NewPadGroupType(2, 40, []int{55, 56, 54, 57, 53, 58, 52, 59, 51, 60, 50, 61, 49, 62, 48, 63, 31, 32, 30, 33, 29, 34, 28, 35, 27, 36, 26, 37, 25, 38, 24, 39, 23, 40, 22, 41, 21, 42, 20, 43, 19, 44, 18, 45, 17, 46, 16, 47, 15, -1, 14, -1, 13, -1, 12, -1, 11, -1, 10, -1, 9, -1, 8, -1, 7, -1, 6, -1, 5, -1, 4, -1, 3, -1, 2, -1, 1, -1, 0, -1}),
@@ -70,53 +42,29 @@ func (seg createSegType9) Build(isBendingPlane bool) CathodeSegmentation {
 				/* Z4 */ NewPadGroupType(3, 40, []int{-1, 16, 15, -1, 17, 14, -1, 18, 13, -1, 19, 12, -1, 20, 11, -1, 21, 10, -1, 22, 9, -1, 23, 8, -1, 24, 7, -1, 25, 6, -1, 26, 5, -1, 27, 4, -1, 28, 3, -1, 29, 2, -1, 30, 1, -1, 31, 0, -1, 48, -1, -1, 49, -1, -1, 50, -1, -1, 51, -1, -1, 52, -1, -1, 53, -1, -1, 54, -1, -1, 55, -1, -1, 56, -1, -1, 57, -1, -1, 58, -1, -1, 59, -1, -1, 60, -1, -1, 61, -1, -1, 62, -1, -1, 63, -1, 47, 32, -1, 46, 33, -1, 45, 34, -1, 44, 35, -1, 43, 36, -1, 42, 37, -1, 41, 38, -1, 40, 39, -1}),
 			},
 			[]padSize{
-				{2.5, 0.5},
 				{5, 0.5},
 			})
 	}
-	return newCathodeSegmentation(9, false,
+	return newCathodeSegmentation(deid, 6, false,
 		[]padGroup{
-			{1035, 3, 0, -45.7142868, -20},
-			{1036, 3, 0, -51.42856979, -20},
-			{1037, 3, 0, -57.1428566, -20},
-			{1038, 3, 0, -62.8571434, -20},
-			{1039, 3, 0, -68.57142639, -20},
-			{1040, 3, 0, -74.2857132, -20},
-			{1041, 3, 0, -80, -20},
-			{1125, 1, 1, 65.7142868, -20},
-			{1126, 4, 1, 54.2857132, -20},
-			{1127, 0, 1, 40, -20},
-			{1133, 1, 1, 25.7142849, -20},
-			{1134, 4, 1, 14.28571415, -20},
-			{1135, 0, 1, 4.440892099e-15, -20},
-			{1141, 1, 1, -14.28571415, -20},
-			{1142, 4, 1, -25.7142849, -20},
-			{1143, 0, 1, -40, -20},
-			{1230, 7, 1, 40, -5},
-			{1231, 5, 1, 48.57143021, -5},
-			{1232, 6, 1, 60, -5},
-			{1233, 8, 1, 70, -5},
-			{1239, 7, 1, -7.105427358e-15, -5},
-			{1240, 5, 1, 8.571428299, -5},
-			{1241, 6, 1, 20, -5},
-			{1242, 8, 1, 30, -5},
-			{1248, 7, 1, -40, -5},
-			{1249, 5, 1, -31.4285717, -5},
-			{1250, 6, 1, -20, -5},
-			{1251, 8, 1, -10, -5},
-			{1325, 2, 0, -80, 0},
-			{1326, 2, 0, -74.2857132, 0},
-			{1327, 2, 0, -68.57142639, 0},
-			{1328, 2, 0, -62.8571434, 0},
-			{1329, 2, 0, -57.1428566, 0},
-			{1330, 2, 0, -51.42856979, 0},
-			{1331, 2, 0, -45.7142868, 0},
+			{1025, 1, 0, 25.7142849, -20},
+			{1026, 2, 0, 14.28571415, -20},
+			{1027, 0, 0, -2.664535259e-15, -20},
+			{1033, 1, 0, -14.28571415, -20},
+			{1034, 2, 0, -25.7142849, -20},
+			{1035, 0, 0, -40, -20},
+			{1130, 5, 0, 0, -5},
+			{1131, 3, 0, 8.571428299, -5},
+			{1132, 4, 0, 20, -5},
+			{1133, 6, 0, 30, -5},
+			{1139, 5, 0, -40, -5},
+			{1140, 3, 0, -31.4285717, -5},
+			{1141, 4, 0, -20, -5},
+			{1142, 6, 0, -10, -5},
 		},
 		[]padGroupType{
 			/* L1 */ NewPadGroupType(20, 4, []int{3, 7, 11, 15, 18, 21, 24, 27, 30, 49, 52, 55, 58, 61, 32, 35, 38, 41, 44, 47, 2, 6, 10, 14, 17, 20, 23, 26, 29, 48, 51, 54, 57, 60, 63, 34, 37, 40, 43, 46, 1, 5, 9, 13, 16, 19, 22, 25, 28, 31, 50, 53, 56, 59, 62, 33, 36, 39, 42, 45, 0, 4, 8, 12, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1}),
 			/* L2 */ NewPadGroupType(20, 4, []int{2, 5, 8, 11, 14, 17, 20, 23, 26, 29, 48, 51, 54, 57, 60, 63, 35, 39, 43, 47, 1, 4, 7, 10, 13, 16, 19, 22, 25, 28, 31, 50, 53, 56, 59, 62, 34, 38, 42, 46, 0, 3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 49, 52, 55, 58, 61, 33, 37, 41, 45, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 32, 36, 40, 44}),
-			/* O1 */ NewPadGroupType(8, 8, []int{40, 32, 56, 48, 24, 16, 8, 0, 41, 33, 57, 49, 25, 17, 9, 1, 42, 34, 58, 50, 26, 18, 10, 2, 43, 35, 59, 51, 27, 19, 11, 3, 44, 36, 60, 52, 28, 20, 12, 4, 45, 37, 61, 53, 29, 21, 13, 5, 46, 38, 62, 54, 30, 22, 14, 6, 47, 39, 63, 55, 31, 23, 15, 7}),
-			/* O2 */ NewPadGroupType(8, 8, []int{7, 15, 23, 31, 55, 63, 39, 47, 6, 14, 22, 30, 54, 62, 38, 46, 5, 13, 21, 29, 53, 61, 37, 45, 4, 12, 20, 28, 52, 60, 36, 44, 3, 11, 19, 27, 51, 59, 35, 43, 2, 10, 18, 26, 50, 58, 34, 42, 1, 9, 17, 25, 49, 57, 33, 41, 0, 8, 16, 24, 48, 56, 32, 40}),
 			/* O3 */ NewPadGroupType(16, 4, []int{3, 7, 11, 15, 19, 23, 27, 31, 51, 55, 59, 63, 35, 39, 43, 47, 2, 6, 10, 14, 18, 22, 26, 30, 50, 54, 58, 62, 34, 38, 42, 46, 1, 5, 9, 13, 17, 21, 25, 29, 49, 53, 57, 61, 33, 37, 41, 45, 0, 4, 8, 12, 16, 20, 24, 28, 48, 52, 56, 60, 32, 36, 40, 44}),
 			/* P1 */ NewPadGroupType(16, 5, []int{47, 46, 41, 36, 63, 58, 53, 48, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 42, 37, 32, 59, 54, 49, 28, 24, 20, 16, 12, 8, 4, 0, -1, -1, 43, 38, 33, 60, 55, 50, 29, 25, 21, 17, 13, 9, 5, 1, -1, -1, 44, 39, 34, 61, 56, 51, 30, 26, 22, 18, 14, 10, 6, 2, -1, -1, 45, 40, 35, 62, 57, 52, 31, 27, 23, 19, 15, 11, 7, 3}),
 			/* P2 */ NewPadGroupType(16, 5, []int{-1, -1, -1, -1, -1, -1, -1, -1, 27, 22, 17, 12, 7, 2, 1, 0, 44, 40, 36, 32, 60, 56, 52, 48, 28, 23, 18, 13, 8, 3, -1, -1, 45, 41, 37, 33, 61, 57, 53, 49, 29, 24, 19, 14, 9, 4, -1, -1, 46, 42, 38, 34, 62, 58, 54, 50, 30, 25, 20, 15, 10, 5, -1, -1, 47, 43, 39, 35, 63, 59, 55, 51, 31, 26, 21, 16, 11, 6, -1, -1}),
@@ -124,11 +72,10 @@ func (seg createSegType9) Build(isBendingPlane bool) CathodeSegmentation {
 			/* Q2 */ NewPadGroupType(14, 5, []int{-1, -1, 35, 62, 57, 52, 31, 26, 21, 16, -1, -1, -1, -1, 44, 40, 36, 63, 58, 53, 48, 27, 22, 17, 12, 8, 4, 0, 45, 41, 37, 32, 59, 54, 49, 28, 23, 18, 13, 9, 5, 1, 46, 42, 38, 33, 60, 55, 50, 29, 24, 19, 14, 10, 6, 2, 47, 43, 39, 34, 61, 56, 51, 30, 25, 20, 15, 11, 7, 3}),
 		},
 		[]padSize{
-			{0.714285714, 2.5},
 			{0.714285714, 5},
 		})
 }
 
 func init() {
-	registerCathodeSegmentationBuilder(9, createSegType9{})
+	mapping.RegisterCathodeSegmentationBuilder(6, createSegType6{})
 }
