+++ mapping/impl4/padgroupgrid.go
@@ -1,4 +1,4 @@
-package mapping
+package impl4
 
 import (
 	"errors"
