+++ mapping/impl4/padgrouptype.go
@@ -1,4 +1,4 @@
-package mapping
+package impl4
 
 import "fmt"
