+++ mapping/impl4/padsize.go
@@ -1,4 +1,4 @@
-package mapping
+package impl4
 
 type padSize struct {
 	x, y float64
