+++ segcontour/bench_test.go
@@ -10,8 +10,8 @@ import (
 
 func getSegs() map[int]mapping.CathodeSegmentation {
 	segs := make(map[int]mapping.CathodeSegmentation)
-	mapping.ForOneDetectionElementOfEachSegmentationType(func(detElemID int) {
-		segs[detElemID] = mapping.NewCathodeSegmentation(detElemID, true)
+	mapping.ForOneDetectionElementOfEachSegmentationType(func(deid int) {
+		segs[deid] = mapping.NewCathodeSegmentation(deid, true)
 	})
 	return segs
 }
@@ -20,8 +20,8 @@ func BenchmarkSegmentationComputeBBoxViaPadLoop(b *testing.B) {
 
 	segs := getSegs()
 
-	for detElemID, seg := range segs {
-		b.Run(strconv.Itoa(detElemID), func(b *testing.B) {
+	for deid, seg := range segs {
+		b.Run(strconv.Itoa(deid), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				mapping.ComputeBBox(seg)
 			}
@@ -33,8 +33,8 @@ func BenchmarkSegmentationComputeBBoxViaContour(b *testing.B) {
 
 	segs := getSegs()
 
-	for detElemID, seg := range segs {
-		b.Run(strconv.Itoa(detElemID), func(b *testing.B) {
+	for deid, seg := range segs {
+		b.Run(strconv.Itoa(deid), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				_ = segcontour.BBox(seg)
 			}
